(#32) Remove business logic from Autosuggest "Begins with" searches
- Remove controller logic for Begins vs. Contains
- Remove Autosuggest begins tests for various combinations of
  Begins and Contains results.
- Consolidate Autosuggest tests.
