+++ src/NCI.OCPL.Api.DrugDictionary/Controllers/AutosuggestController.cs
@@ -93,54 +93,7 @@ namespace NCI.OCPL.Api.DrugDictionary.Controllers
             // searchText uses a catch-all route, make sure it's been decoded.
             searchText = WebUtility.UrlDecode(searchText);
 
-            switch (matchType)
-            {
-                case MatchType.Begins:
-                    return await DoBegins(searchText, size, includeResourceTypes, includeNameTypes, excludeNameTypes);
-
-                case MatchType.Contains:
-                    return await _autosuggestQueryService.GetSuggestions(searchText, matchType, size, includeResourceTypes, includeNameTypes, excludeNameTypes);
-
-                default:
-                    throw new APIErrorException(400, $"Unknown `matchType` parameter: '{matchType}'.");
-            }
-        }
-
-        /// <summary>
-        /// Helper method to encapsulate the logic for a Begins autosuggest search.
-        /// </summary>
-        /// <param name="searchText">Text to match against</param>
-        /// <param name="size">The number of records to retrieve.</param>
-        /// <param name="includeResourceTypes">The DrugResourceTypes to include. Default: All</param>
-        /// <param name="includeNameTypes">The name types to include. Default: All</param>
-        /// <param name="excludeNameTypes">The name types to exclude. Default: All</param>
-        /// <returns>A collection of up to size number of Suggestion objects, combining the results from
-        /// a Begins and a Contains search.</returns>
-        private async Task<Suggestion[]> DoBegins(string searchText, int size,
-            DrugResourceType[] includeResourceTypes,
-                TermNameType[] includeNameTypes,
-                TermNameType[] excludeNameTypes
-        )
-        {
-            // Autosuggest for Begins actually combines both a Beginsand a Contains search,
-            // returning a maximum of size results.
-            Task<Suggestion[]>[] queries = new Task<Suggestion[]>[]
-            {
-                _autosuggestQueryService.GetSuggestions(searchText, MatchType.Begins, size, includeResourceTypes, includeNameTypes, excludeNameTypes),
-                _autosuggestQueryService.GetSuggestions(searchText, MatchType.Contains, size, includeResourceTypes, includeNameTypes, excludeNameTypes)
-            };
-
-            // An array of arrays.
-            Suggestion[][] suggestions = await Task.WhenAll(queries);
-
-            // suggstions[0] is the begins result.
-            // suggestions[1] is the contains result.
-            List<Suggestion> results = new List<Suggestion>(suggestions[0]);
-            results.AddRange(suggestions[1]);
-
-            // Return no more than the requested number of results.
-            int returnSize = (size < results.Count ? size : results.Count);
-            return results.GetRange(0, returnSize).ToArray();
+            return await _autosuggestQueryService.GetSuggestions(searchText, matchType, size, includeResourceTypes, includeNameTypes, excludeNameTypes);
         }
     }
 }
